JAMES-2586 Rename Postgres Subscription
